CSS updates for spacing and color
