+++ src/App.css
@@ -6,7 +6,15 @@
 
 .App header {
   text-align: center;
-  color: #f2f2f2;
+  color: var(--my-white);
+}
+
+.loading {
+  margin-top: 25px;
+}
+
+.loading p { 
+  text-align: center;
 }
 
 .loading-spinner {
@@ -14,15 +22,13 @@
   justify-content: center;
 }
 
+.dashboard {
+  margin: 10px;
+}
 
-/* 
-h1, h2 {
-  color: white;
-} */
-
-/* ul {
-  list-style-type: none;
-} */
+.analysis h1 {
+  text-align: center;
+}
 
 @media only screen and (min-width: 640px) {
   .analysis {
