+++ src/components/stockSearch.css
@@ -2,12 +2,41 @@
   text-align: center;
 }
 
-.stock-search input {
-
+.stock-search input[type="text"] {
+  width: 80%;
+  max-width: 400px;
+  border: 4px solid var(--my-orange);
 }
 
-.stock-search input[type="search"] {
-  width: 80%;
-  max-width: 300px;
-  font-size: 50px;
+.stock-search input[type="text"], .stock-search button {
+  font-size: 30px;
+  height: 40px;
+  padding: 5px;
+  border-radius: 5px;
+  box-sizing: content-box;
+  margin-bottom: 15px;
+}
+
+.stock-search button {
+  border: 4px solid var(--my-orange);
+  color: var(--my-blue);
+  background-color: var(--my-orange);
+}
+
+.stock-search button:hover {
+  color: var(--my-orange);
+  background-color: var(--my-blue);
+}
+
+@media only screen and (min-width: 640px) {
+  .stock-search input[type="text"] {
+    margin-right: 20px;
+    font-size: 40px;
+    height: 50px;
+  }
+
+  .stock-search input[type="text"], .stock-search button {
+    font-size: 40px;
+    height: 50px;
+  }
 }
