+++ src/components/tweets.css
@@ -26,9 +26,9 @@
   /* font-weight: bold; */
   line-height: 16px;
   border-bottom: 2px solid black;
-  /* margin: 10px 5px; */
+  margin: 10px;
   /* padding: 0 16px 16px 16px; */
-  width: 100%;
+  /* width: 100%; */
   /* min-width: 400px; */
 }
 
@@ -48,7 +48,7 @@
 .tweet-username {
   display: flex;
   align-items: flex-end;
-  padding-left: 10px;
+  /* padding-left: 10px; */
 }
 
 ul {
