+++ src/index.css
@@ -1,6 +1,13 @@
+:root {
+  --my-white: #f2f2f2;
+  --my-blue: #4382c3;
+  --my-green: #45B25C;
+  --my-orange: #F4C01E;
+}
+
 body {
   margin: 0;
   padding: 0;
   font-family: 'Lato', sans-serif;
-  background-color: #4382c3;
+  background-color: var(--my-blue);
 }
